Added more casts in nrSteps to make it work with unsigned types.
git-svn-id: https://svn.opendtect.org/od/trunk@30605 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Basic/ranges.h
@@ -735,7 +735,7 @@ template <class T>
 inline int StepInterval<T>::nrSteps() const
 {
     if ( !step ) return 0;
-    int ret = (((int)this->start) - this->stop) / step;
+    int ret = (((int)this->start) - ((int) this->stop)) / ((int) step);
     return ret < 0 ? -ret : ret;
 }
